Migrate to Swift 4, add proprer support for iPhone X
