+++ Sources/Instructions/Instructions.swift
@@ -26,6 +26,5 @@ struct Constants {
     static let overlayFadeAnimationDuration: TimeInterval = 0.3
     static let coachMarkFadeAnimationDuration: TimeInterval = 0.3
 
-    //swiftlint:disable line_length
     static let overlayColor = #colorLiteral(red: 0.9086670876, green: 0.908688426, blue: 0.9086769819, alpha: 0.65)
 }
