+++ Sources/Instructions/Protocols/Public/CoachMarkArrowView.swift
@@ -23,9 +23,9 @@
 import UIKit
 
 /// A protocol to which all the "arrow views" of a coach mark must conform.
-public protocol CoachMarkArrowView : class {
+public protocol CoachMarkArrowView: class {
     /// A method to change the arrow highlighted state.
     /// If you feel the arrow should mirror the state of the "body view",
     /// You will most likely change the background color of the view here.
-    var isHighlighted: Bool { set get }
+    var isHighlighted: Bool { get set }
 }
