+++ Sources/Instructions/Protocols/Public/CoachMarkSkipView.swift
@@ -23,7 +23,7 @@
 import UIKit
 
 /// A protocol to which all the "skip views" must conform.
-public protocol CoachMarkSkipView : class {
+public protocol CoachMarkSkipView: class {
     /// The control that will trigger the stop, in the display flow.
     var skipControl: UIControl? { get }
     var asView: UIView? { get }
